final/resume_based_job_recommendation-/requirements.txt
dotenv
pymongo
fastapi
uvicorn
langchain_groq
langchain_community
langchain
pdfminer.six
langchain_aws
agno
pypdf
duckduckgo-search
langgraph
markdownify
markdown
transformers
torch
passlib[bcrypt]
email-validator
pandas